SO-4731: Move custom Enums to the model package
