--- com.b2international.snomed.ecl/src/com/b2international/snomed/ecl/Domain.java
@@ -1,26 +1,0 @@
-/*
- * Copyright 2021 B2i Healthcare Pte Ltd, http://b2i.sg
- * 
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * 
- *      http://www.apache.org/licenses/LICENSE-2.0
- * 
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.b2international.snomed.ecl;
-
-/**
- * Enumerates the "subject" of the filter expression it operates on.
- * 
- * @since 1.5
- */
-public enum Domain {
-	CONCEPT,
-	DESCRIPTION;
-}
